Add Zprobe(AutoleveL) in Configuration menu
Add Zprobe function from menu, proposing/asking user to clean nozzle and
plate first then heating bed and extruder then doing Zprobe, when done
it propose user to save result
Unlike G32 command, this function handle failure and put nozzle in safe
position what ever happen to avoid any hit on plates, this function does
not take in account Zprobe end script but only Zprobe start script,
anyway both are are empty for DaVinci so no real impact.
Same for Z_PROBE_REPETITIONS it is not taken in account but it is set to
1 for DaVinci so not impact neither, failure point is not specified -
this can be added later.